Don't assume opaque hipDevicePtr_t type is an int
CUdeviceptr is documented<https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TYPES.html#group__CUDA__TYPES_1g183f7b0d8ad008ea2a5fd552537ace4e>
as "an unsigned integer type whose size matches the size of a pointer
on the target platform"

The opaque hipDevicePtr_t type is a void *.

The difference in implementation of these opaque types is that there
are subtle differences in how they are treated by the language.
Casting away const qualifiers is allowed for int, but not void *.

The base_region_t type is templated and only allows void and const void
but since the implementation of address_t is actually void, it turned
out to give rise to a duplicate signature of the ctor after
instantiating base_region_t.